ocr-server/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
easyocr==1.7.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.9.0
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.1
ninja==1.11.1.1
numpy==2.1.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==11.0.0
pyclipper==1.3.0.post6
PyPDF2==3.0.1
python-bidi==0.6.3
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
setuptools==75.2.0
shapely==2.0.6
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchvision==0.20.0
typing_extensions==4.12.2
Werkzeug==3.0.4
